EquiTrack/requirements/base.txt
Django==1.8.9
django-braces==1.8.1
django-model-utils==2.3.1
logutils==0.3.3
psycopg2==2.6.1
dj-database-url==0.3.0
django-celery==3.1.17
django-celery-email==1.1.3
django-filer==1.0.5
#django-storages-redux==1.3
django-suit==0.2.16
django-suit-ckeditor==0.0.2
django-reversion==1.10.1
django-import-export==0.3.1
django-autocomplete-light==2.2.10
django-generic-links==0.4.0
jsonfield==1.0.3
django-logentry-admin==0.1.5
djrill==2.0
django-smart-selects==1.2.2
djangorestframework==3.3.2
djangorestframework-csv==1.3.3
djangorestframework-jwt==1.7.2
djangorestframework-hstore==1.3
django-rest-swagger==0.3.5
django-redis-cache==1.6.4
hiredis==0.2.0
GDAL==1.10.0
# Fiona==1.5.0
# Shapely==1.5.7
cartodb==0.8.1
gunicorn==19.4.1
django-datetime-widget==0.9.3
django-leaflet==0.17.0
django-geojson==2.9.0
django-post_office==2.0.5
django-cors-headers==1.1.0
django-mathfilters==0.3.0
django-tenant-schemas==1.5.7
django-allauth==0.24.1
tenant-schemas-celery==0.1.5
drf-nested-routers==0.11.1
azure-storage==0.20.2
pandas==0.17.1
requests==2.8.1
django-analytical==2.0.0
xlrd==0.9.3
dj-static==0.0.6
django-easy-pdf
xhtml2pdf
reportlab

# testing libs
coverage==4.0
factory_boy==2.5.2
django-debug-toolbar==1.4

git+https://github.com/robertavram/djangosaml2.git#egg=djangosaml2
git+https://github.com/robertavram/django-storages.git#egg=django-storages
git+https://github.com/jamescw/django-paintstore.git
git+https://github.com/UNICEFLebanonInnovation/ActvityInfoPython.git

